move types to common
