+++ common/types/formula.ts
@@ -1,0 +1,11 @@
+export enum Formula {
+  Manual = 'MANUAL',
+  WeightedAverage = 'WEIGHTED_AVERAGE',
+}
+
+export interface FormulaPreview {
+  id: Formula;
+  name: string;
+  attributes: Array<string>;
+  codeSnippet: string;
+}
