+++ common/types/index.ts
@@ -5,5 +5,6 @@
 export * from './attainment';
 export * from './auth';
 export * from './course';
+export * from './formula';
 export * from './language';
 export * from './user';
